refactor(react): modularize components
